requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
python-dotenv==1.0.1
PyJWT==2.10.1
python-socketio==5.12.1
python-multipart==0.0.20
firebase-admin==6.6.0
httpx==0.28.1
pydantic==2.10.4
